% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFuncs.R
\name{myCube}
\alias{myCube}
\title{My wonderful cube function}
\usage{
myCube(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the cube (third power) of the numeric vector x
}
\description{
My wonderful cube function
}
\examples{
myCube(1:9)
set.seed(1234)
myCube(rnorm(10))
}
