% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFuncs.R
\name{mySquare}
\alias{mySquare}
\title{My wonderful square function}
\usage{
mySquare(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the square of the numeric vector x
}
\description{
My wonderful square function
}
\examples{
mySquare(1:9)
set.seed(1234)
mySquare(rnorm(10))
}
